fix all builtins and cd works on the first case
