+++ src/misc_utils.c
@@ -6,7 +6,7 @@
 /*   By: bschwell <student@42.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/27 11:32:18 by bschwell          #+#    #+#             */
-/*   Updated: 2024/12/27 12:23:15 by bschwell         ###   ########.fr       */
+/*   Updated: 2024/12/31 17:25:42 by bschwell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
